+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/entity/Parent.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2014 - 2019 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.testsuite.entity;
+
+import javax.persistence.Column;
+import javax.persistence.DiscriminatorColumn;
+import javax.persistence.DiscriminatorType;
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.persistence.Table;
+import java.io.Serializable;
+
+/**
+ *
+ * @author Christian Beikov
+ * @since 1.4.0
+ */
+@Entity
+@Table(name = "PARENT_TBL")
+@DiscriminatorColumn(name = "DTYPE", discriminatorType = DiscriminatorType.INTEGER)
+@DiscriminatorValue("0")
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+public class Parent implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    private Long id;
+    private Integer number;
+
+    @Id
+    @GeneratedValue
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    @Column(name = "order_nr")
+    public Integer getNumber() {
+        return number;
+    }
+
+    public void setNumber(Integer number) {
+        this.number = number;
+    }
+
+}